Some minor refactoring and adding columns ordering feature
